clean up pixeliation.mesh.test_delaunay
