+++ test_autoarray/inversion/pixelization/data_grid/test_kmeans.py
@@ -4,6 +4,68 @@ import pytest
 import autoarray as aa
 
 
+def test__brightness__weight_map_from():
+    adapt_data = np.array([0.0, 1.0, 0.0])
+
+    pixelization = aa.image_mesh.KMeans(
+        pixels=5, weight_floor=0.0, weight_power=0.0
+    )
+
+    weight_map = pixelization.weight_map_from(adapt_data=adapt_data)
+
+    assert (weight_map == np.ones(3)).all()
+
+    pixelization = aa.image_mesh.KMeans(
+        pixels=5, weight_floor=0.0, weight_power=1.0
+    )
+
+    weight_map = pixelization.weight_map_from(adapt_data=adapt_data)
+
+    assert (weight_map == np.array([0.0, 1.0, 0.0])).all()
+
+    pixelization = aa.image_mesh.KMeans(
+        pixels=5, weight_floor=1.0, weight_power=1.0
+    )
+
+    weight_map = pixelization.weight_map_from(adapt_data=adapt_data)
+
+    assert (weight_map == np.array([1.0, 2.0, 1.0])).all()
+
+    pixelization = aa.image_mesh.KMeans(
+        pixels=5, weight_floor=1.0, weight_power=2.0
+    )
+
+    weight_map = pixelization.weight_map_from(adapt_data=adapt_data)
+
+    assert (weight_map == np.array([1.0, 4.0, 1.0])).all()
+
+    adapt_data = np.array([-1.0, 1.0, 3.0])
+
+    pixelization = aa.image_mesh.KMeans(
+        pixels=5, weight_floor=0.0, weight_power=1.0
+    )
+
+    weight_map = pixelization.weight_map_from(adapt_data=adapt_data)
+
+    assert (weight_map == np.array([0.0, 0.5, 1.0])).all()
+
+    pixelization = aa.image_mesh.KMeans(
+        pixels=5, weight_floor=0.0, weight_power=2.0
+    )
+
+    weight_map = pixelization.weight_map_from(adapt_data=adapt_data)
+
+    assert (weight_map == np.array([0.0, 0.25, 1.0])).all()
+
+    pixelization = aa.image_mesh.KMeans(
+        pixels=5, weight_floor=1.0, weight_power=1.0
+    )
+
+    weight_map = pixelization.weight_map_from(adapt_data=adapt_data)
+
+    assert (weight_map == np.array([3.0, 3.5, 4.0])).all()
+
+
 def test__from_pixels_grid_and_weight_map():
     mask = aa.Mask2D(
         mask=np.array(
